pyproject.toml
[tool.poetry]
name = "healthy-habits-tracker"
version = "0.1.0"
description = ""
authors = ["Alexandr Abramov <https://github.com/Abramov0Alexandr>"]
readme = "README.md"
packages = [{include = "healthy_habits_tracker"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.4"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
ipython = "^8.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.2.0"
celery = "^5.3.4"
redis = "^5.0.0"
eventlet = "^0.33.3"
django-celery-beat = "^2.5.0"
notifiers = "^1.3.3"
coverage = "^7.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
